Changed about everything to NavigationLink with NavigationDestinations, but Listview Broken
Ugh, I tried, and tried, and even ChatGPT wasn't much help. But it's good enough for now, I guess. This isn't a real app, but I would like to come back to this and figure out why the ListView always brings up another List View after you navigate to a mission or astronaut.